+++ src/assinafy/__init__.py
@@ -25,7 +25,7 @@ from .types import (
     WebhookEventType,
 )
 
-__version__ = "1.3.2"
+__version__ = "1.4.0"
 
 __all__ = [
     "ApiError",